requirements.txt
PySide6>=6.5.0
discord.py>=2.3.2
psutil>=5.9.0
requests>=2.31.0